+++ changelog.md
@@ -5,6 +5,7 @@
 ### New Features
 - dwd/profile/mrmirz-4412 redirects to profile/mrmirz
 - rejected posts are no longer deleted and display rejection notice for resubmission guidance
+- Option to be notified for new posts and when a post is marked as stable
 
 ### Improvements
 - profile and post name on detail pages now wraps and dynamically resizes